# gudangku/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GudangKu",
        "description": "API Documentation for GudangKu",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/update_telegram_id": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update telegram token id",
                "operationId": "b3319a420453e08af500d97aa4d48b6d",
                "responses": {
                    "200": {
                        "description": "telegram id updated! and validation has been sended to you"
                    },
                    "404": {
                        "description": "telegram id failed to update"
                    },
                    "409": {
                        "description": "telegram user id has been used"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/register/token": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Check and send validation token to the user who in registration process",
                "operationId": "f9a08af4b40b9ac720ec6222bb337c0f",
                "responses": {
                    "200": {
                        "description": "the validation token has been sended to {email} email account"
                    },
                    "409": {
                        "description": "there already a request with same username / username already being used"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/register/account": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register account and accept validation",
                "operationId": "29226b21d6a5a79d2e87ade3dfed667f",
                "responses": {
                    "200": {
                        "description": "account is registered"
                    },
                    "404": {
                        "description": "Token is invalid"
                    },
                    "409": {
                        "description": "username already used"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/register/regen_token": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Regenerate registration token",
                "operationId": "de83a252f4cbd7f245743ba8e36c2da2",
                "responses": {
                    "200": {
                        "description": "the validation token has been sended to {email} email account"
                    },
                    "404": {
                        "description": "request not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/user/update_timezone_fcm": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update user timezone",
                "operationId": "a8507903e83d7a4e0cc32bc26c47889f",
                "responses": {
                    "200": {
                        "description": "firebase message token / timezone has been updated"
                    },
                    "422": {
                        "description": "Timezone is invalid"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}